backend/requirements.txt
fastapi==0.85.1
torch==2.4.0
langchain-core==0.2.24
langchain-community==0.2.10
pypdf==4.3.1
chromadb==0.3.29
sentence-transformers==2.2.2
python-dotenv==1.0.1
InstructorEmbedding==1.0.1
pyproject-toml==0.0.10
python-multipart==0.0.9
uvicorn==0.30.3
langchain_experimental
faiss-cpu
pymupdf